refactor(folderStructure): move models outside of store
