+++ src/models/SliceStatus.ts
