+++ src/models/TrackerEntry.ts
