+++ src/models/TrackerStatus.ts
